Added-Added delete content when thing is deleted. Also improved error handling.

+++ thing/src/main/java/com/fkhr/thingsorganizer/thing/service/ThingServiceImpl.java
@@ -2,8 +2,10 @@ package com.fkhr.thingsorganizer.thing.service;
 
 import com.fkhr.thingsorganizer.common.exeptionhandling.CustomError;
 import com.fkhr.thingsorganizer.common.exeptionhandling.CustomExeption;
+import com.fkhr.thingsorganizer.common.util.EntityType;
 import com.fkhr.thingsorganizer.thing.model.Container;
 import com.fkhr.thingsorganizer.thing.model.Thing;
+import com.fkhr.thingsorganizer.thing.proxy.ContentProxy;
 import com.fkhr.thingsorganizer.thing.repository.ThingRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,6 +13,7 @@ import org.springframework.context.annotation.Lazy;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -20,11 +23,14 @@ import java.util.Optional;
 public class ThingServiceImpl implements ThingService {
     final private ThingRepository thingRepository;
     final private ContainerService containerService;
-
+    final private ContentProxy contentProxy;
     @Autowired
-    public ThingServiceImpl(ThingRepository thingRepository, @Lazy ContainerService containerService) {
+    public ThingServiceImpl(ThingRepository thingRepository,
+                            @Lazy ContainerService containerService,
+                            ContentProxy contentProxy) {
         this.thingRepository = thingRepository;
         this.containerService = containerService;
+        this.contentProxy = contentProxy;
     }
 
     @Override
@@ -41,7 +47,7 @@ public class ThingServiceImpl implements ThingService {
         else{
             Thing existingThing = this.load(thing.getId());
             if(existingThing == null){
-                throw new CustomExeption(CustomError.THING_NOT_FOUND, HttpStatus.NOT_FOUND);
+                throw new CustomExeption(CustomError.THING_NOT_FOUND, HttpStatus.ACCEPTED);
             }
             else {
                 result = thingRepository.save(thing);
@@ -57,17 +63,25 @@ public class ThingServiceImpl implements ThingService {
     public void delete(Long id){
         Thing thing = this.load(id);
         if(thing != null) {
-            thingRepository.delete(thing);
+            ResponseEntity deleteContentResponse = contentProxy.deleteContentByOwner(EntityType.THING, id.toString());
+
+            if(deleteContentResponse.getStatusCode().value() == 200 ||
+                    deleteContentResponse.getStatusCode().value() == 204) {
+                thingRepository.delete(thing);
+            }
+            else{
+                throw new CustomExeption(CustomError.THING_DELETE_CONTENT_NOT_DELETED, HttpStatus.ACCEPTED);
+            }
         }
         else
-            throw new CustomExeption(CustomError.THING_NOT_FOUND, HttpStatus.NOT_FOUND);
+            throw new CustomExeption(CustomError.THING_NOT_FOUND, HttpStatus.ACCEPTED);
     }
 
     @Override
     public Thing load(Long id) {
         Optional<Thing> result = thingRepository.findById(id);
         if(result.equals(Optional.empty())){
-            throw new CustomExeption(CustomError.THING_NOT_FOUND, HttpStatus.NOT_FOUND);
+            throw new CustomExeption(CustomError.THING_NOT_FOUND, HttpStatus.ACCEPTED);
         }else{
             return result.get();
         }
@@ -102,7 +116,7 @@ public class ThingServiceImpl implements ThingService {
             return thingRepository.updateContainer(oldContainerId, newContainerId);
         }
         else{
-            throw new CustomExeption(CustomError.CONTAINER_NOT_FOUND, HttpStatus.NOT_FOUND);
+            throw new CustomExeption(CustomError.CONTAINER_NOT_FOUND, HttpStatus.ACCEPTED);
         }
     }
 }
